requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
decorator==4.4.2
einops==0.2.0
filelock==3.0.12
gdown==3.11.0
idna==3.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
lmdb==1.0.0
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
prefetch-generator==1.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.2
scipy==1.7.0
Shapely==1.7.1
six==1.16.0
tifffile==2021.7.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
vietocr==0.3.5
